requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
aiogram>=3.0
python-dotenv
sqlalchemy
asyncpg
pgvector
sentence-transformers
groq
aiofiles